build_module: Add symvers/intermediates copies
Currently, the external module build with DDK does not produce the
Modules.symvers or intermediate files in the correct locations. All
of the files are just generated up-front.

To maintain backwards-compatibility, copy the files to the expected
locations.

Eventually, once all modules have transitioned to DDK, these copies
will not be needed.

Change-Id: Id825164bb29a9561e098527fb4f28dce4c054e37
Signed-off-by: John Moon <quic_johmoo@quicinc.com>