build_module: Only concatenate symvers from same variant
Currently, the Modules.symvers file for a given package is generated
by concatenating all of the *_Modules.symvers generated in the dir
directory. In some cases, multiple builds of different variants will
run one after another which causes module duplication in the output
file. This leads to errors like this:

ERROR: modpost: <package>: '<symbol>' exported twice. Previous export was in <package>.ko

Narrow down the concatenation to the symvers files which match the
target/variant combo to avoid this duplication.

Change-Id: I7640b28615aef2473dc9848ab3d53c6f08215478
Signed-off-by: John Moon <quic_johmoo@quicinc.com>