kleaf: Enable incremental-fs for 16kB page size targets
The Incremental File System (IncFS) kernel module enables the Android
OS to receive streamed APKs over the Android Debug Bridge (ADB). It
creates a new virtual file system that sits on top of the existing
Android file system. This complements changes in the framework and SDK
to enable app and game developers to deploy large APKs through the ADB.

Bug: 338659380
Test: atest IncrementalFeatureTest on ARM64 Cuttlefish with 16 KB page size build.
Change-Id: Ib06dcf966e9fe012b9e8bbbd63019d4c2070a022
Signed-off-by: Vilas Bhat <vilasbhat@google.com>