kleaf: kernel_config split post_defconfig_fragments
... into two lists: post_defconfig_fragments_inherited, and
post_defconfig_fragments_non_inherited. This allows us to treat them
differently later.

The defconfig fragment for trim_nonlisted_kmi (CONFIG_TRIM_UNUSED_KSYMS)
is NOT inherited because for the device kernel_build(), this only refers
to the KMI surface exposed by the unsigned in-tree modules to external
modules. Before and after this change, CONFIG_TRIM_UNUSED_KSYMS is not
enabled by default for building in-tree unsigned modules.

Bug: 416803611
Change-Id: I5e543652581f36228b0f336caf24baff7ae48640