kleaf: Add support for building ditto suite
This introduces a new config `ditto` to build the ditto suite, it
requires a special configuration as it depends on two repos that are not
part of the standard checkout.

This add a new fake module for repos which don't include the
test/dittuosuite project, as it's not available in any registry.

Tested: tools/bazel build @dittosuite//:all --config=ditto
Tested: tools/bazel build //common:kernel_aarch64
Bug: 429971301
Change-Id: I30a6191a40cc8b3bb566e9c73ec194683081acdf