kleaf: vendor_boot_image pass FilesToRunProvider for mkbootimg to subrules.
A bug in Bazel is that, if the ctx.executable.foo comes from the
rule attribute, it is a File. When this File is passed to
subrule_ctx.actions.run(tools=) in a subrule, its runfiles are not
automatically added, unlike ctx.actions.run(tools=) directly in the
rule.

Hence, when passing the executable to subrules, we need to pass the
FilesToRunProvider instead.

Bug: 433545832
Link: https://github.com/bazelbuild/bazel/issues/26627
Change-Id: I7423c9d18b73cc9f551aabaf234d3ebe7c0f1116