GKI: kernel_processing: also return the banner string
The function get_linux_banner can find a Linux banner inside an
arbitrary byte string (typically a kernel image).

It now also extracts and returns the banner itself (less the trailing
newline and NUL terminator) to avoid having to duplicate this
functionality in its callers.

Also no longer include the whole of the input in exceptions.

Bug: 393559289
Change-Id: I6298eecf9d6bd16461b556cd6a9a4aed749f4483
Signed-off-by: Giuliano Procida <gprocida@google.com>