kleaf: auto-load build.config.constants, even with build_config.
Previously, if kernel_build.build_config is set, for backwards
compatibility, it was required that the build config must start with

   source ${ROOT_DIR}/${KERNEL_DIR}/build.config.constants

Otherwise, build.config.constants would not be loaded.

This design not only made migrations for partners harder, but also
impose a challenge when we try to eliminate build.config.constants.

After this change, when
--incompatible_autoload_build_config_constants=always,
it always load build.config.constants regardless of
whether build_config is set. In addition, to catch overridden values
that were previously allowed unintentionally, it comes with a check
after `source build_config` that the values must still be the same.

In other words, with
--incompatible_autoload_build_config_constants=always:

- kernel_build with build_config no longer need build.config.constants
  as the first line.
- If a kernel_build has build_config that loads different
  build.config.constants, it will no longer be allowed.

This behavior is controlled by the
--incompatible_autoload_build_config_constants
flag. The new behavior sets the flag value to "always".
The old behavior sets the flag value to "if_unset".

Bug: 448148817
Bug: 236012223
Change-Id: I0d7992ccc6e2f2326a065e90e0d6d89825bdf1f6