Kleaf: Add avb test repo with minimal build description
* This change allows usage of the test keys inside the avb
repository instead of the copy from prebuilts.

* This change is a cleanup to have only one source of truth
and to reduce file duplications.

* This change introduces  a hard dependency on avb repository
which might break users who do not include avb in their manifests,
in such case, please refer to http://r.android.com/3580519 for
an example on how to add it.

* The entire @avb repository is not used as it currently conflicts
with GBL usage of a repo with the same name.

Bug: 447314130
Change-Id: I61a60c11ed5bc55b47e0ff24c3d2df6d41634a7f