kleaf: integration_test: override kleaf_repo_dir for wrapper.
If the wrapper is invoked not via bazel.sh but via integration test,
its path looks like:

<real workspace>/out/bazel/output_user_root/9299907c98d95b1c45586a015ae60386/execroot/_main/bazel-out/k8-fastbuild/bin/build/kernel/kleaf/tests/integration_test/integration_test.sh.runfiles/_main/build/kernel/kleaf/wrapper

Depending on the test case, the wrapper may be invoked on different
workspace roots, some of them may be mount roots.

To properly infer the value of kleaf_repo_dir in order to determine
whether _kleaf_repository_is_top_workspace(), we have to override
the value of kleaf_repo_dir in the integration test.

Bug: 413506174
Change-Id: Id8ea85eeacdd7017ee9262402cc1ead3744d1f53