kleaf: Add rules_rust & rules_rust_tinyjson.
rules_rust could technically be a dev_dependency. However, we have
announced to add rules_rust a long time ago, and it is well expected
that eventually we'll use rust toolchain to build our kernel_build as
well. To avoid making ddkv2 setup complicated, just mark it as
a non-dev_dependency.

Bug: 328465178
Change-Id: I9ba3092a22235d5e52182c3dd1139b7a42fb9bb9