kleaf: remove excludes from rsync command
When syncing the modules install staging directly, the rsync command
excludes "build" and "source" folders as it assumes they are symbolic
links. However, this causes issues when a driver is under a folder with
one of those names as those drivers then get excluded.

In my analysis of the <name>_modules_install folder, I didn't find any
build and source symbolic links. So let's remove that logic to fix this.

Bug: 456197602
Change-Id: I5a8036e8749a89dcde1b46616154247660c21787
Signed-off-by: Will McVicker <willmcvicker@google.com>