prepare_vendor: Remove "flatten" version of GKI modules.
Starting with the android14-6.1 kernel and later, the
system_dlkm_staging_archive.tar.gz file contains original and "flatten"
versions of the compiled modules directory.
The "flatten" version is not used in the gen3gvmcoqos target. However,
due to hard-coded logic in the core scripts, modules from this folder and
subfolders are copied to the target system_dlkm image.
Hence, delete unnecessary copies of modules.

Change-Id: Id8e7c4d984846004d08708a8204c84d1d8692eb8
Signed-off-by: Andrii Smolskyi <asmolsky@qti.qualcomm.com>